Gracefully handle non-GUID for DELETE
Again, it may seem like an odd choice to accept a malformed ID and then
just do nothing, but like the argument put forth for the GET
implementation, a URL like /reservations/bar represents a resource that
doesn't exist. You'd think, then, that a more appropriate response would
be 404 Not Found.

DELETE operations, however, are supposed to be idempotent, so even for a
valid URL like /reservations/79f53e9d9a66458ab79e11da130bf1d8, the first
time you DELETE it, the response should be 200 OK. If you do it again,
the resource IS NO LONGER there, but due to idempotency, the response
ought still to be 200 OK.

To generalise, I think that this implies that a DELETE against an absent
resource should just be a successful no-op.

+++ Restaurant.RestApi/ReservationsController.cs
@@ -95,8 +95,8 @@ namespace Ploeh.Samples.Restaurant.RestApi
         [HttpDelete("{id}")]
         public async Task Delete(string id)
         {
-            var rid = new Guid(id);
-            await Repository.Delete(rid).ConfigureAwait(false);
+            if (Guid.TryParse(id, out var rid))
+               await Repository.Delete(rid).ConfigureAwait(false);
         }
     }
 }
